+++ src/common.sh
@@ -22,6 +22,8 @@
 #######################################################################
 
 COMMON_EFI_FILE="/sys/firmware/efi"
+COMMON_ESP_PATH="/boot/efi"
+export COMMON_EFI_PATH="EFI/Linux"
 
 ###
 # Check if the system use EFI
@@ -32,4 +34,84 @@ COMMON_EFI_FILE="/sys/firmware/efi"
 ###
 common_is_efi_system() {
     [ -d "${COMMON_EFI_FILE}" ]
+}
+
+###
+# Get the device name from the mounted directory
+# ARGUMENTS:
+#   1 - Directory mounted (/ or /boot/efi)
+# OUTPUTS:
+#   UUID part
+# RETURN:
+#   none
+###
+common_get_dev_name() {
+    df -h "$1" | tail -1 | cut -d ' ' -f1
+}
+
+###
+# Get the UUID partition of a directory mounted
+# ARGUMENTS:
+#   1 - Device (Call common_get_dev_name)
+# OUTPUTS:
+#   UUID part
+# RETURN:
+#   none
+###
+common_get_dev_uuid() {
+    blkid "$1" | sed -e 's|.* UUID="\(.*\)|\1|' | sed 's|" .*||'
+}
+
+###
+# Get the Available space of a partition
+# ARGUMENTS:
+#   1 - Device (Call common_get_dev_name)
+# OUTPUTS:
+#   Available space
+# RETURN:
+#   none
+###
+common_get_dev_avail() {
+    df --block-size="1M" --output="avail" "$1" | tail -1 | tr -d ' '
+}
+
+###
+# If not, install the uki in the efi boot partition
+# ARGUMENTS
+#   1 - uki path
+#   2 - efi dir
+# OUTPUTS:
+#   Status
+# RETURN:
+#   exit 2 in failure
+###
+common_install_uki_in_efi() {
+    uki_path="$1"
+    image=$(basename "$1")
+    efi_d="$2"
+    esp_efi_d="${COMMON_ESP_PATH}/${efi_d}"
+    # uki_path="/usr/lib/modules/${kerver}/${image}"
+    [ ! -d "${esp_efi_d}" ] && mkdir -p "${esp_efi_d}"
+    if [ ! -f "${esp_efi_d}/${image}" ]; then
+        if [ ! -f "${uki_path}" ]; then
+            echo_error "Unable to find the UKI file: ${uki_path}"
+            exit 2
+        else
+            echo_debug "Install UKI in ${esp_efi_d}/${image}"
+            efi_dev="$(common_get_dev_name ${COMMON_ESP_PATH})"
+            uki_size="$(du -m0 "$uki_path" | cut -f 1)"
+            efi_avail="$(common_get_dev_avail "$efi_dev")"
+            echo_debug "${efi_avail}M available on efi partition"
+            echo_debug "Size of uki file: ${uki_size}M"
+            if [ "$uki_size" -gt "$efi_avail" ]; then
+                echo_error "No space left on efi partition to install uki"
+                echo_error "Need ${uki_size}M, Available: ${efi_avail}M"
+                exit 2
+            fi
+            if ! cp "$uki_path" "${esp_efi_d}/${image}"; then
+                echo_error "Error when installing ${esp_efi_d}/${image}"
+                exit 2
+            fi
+        fi
+    fi
 }